chore(python): move types into a subfolder
