+++ python/bullmq/types/queue_options.py
@@ -1,0 +1,13 @@
+
+from typing import TypedDict
+
+
+class QueueOptions(TypedDict, total=False):
+    """
+    Options for the Queue class.
+    """
+
+    prefix: str
+    """
+    Prefix for all queue keys.
+    """
