+++ python/bullmq/types/retry_job_options.py
@@ -1,0 +1,8 @@
+
+from typing import TypedDict
+
+
+class RetryJobsOptions(TypedDict, total=False):
+    state: str
+    count: int
+    timestamp: int
